# project_atlas_openapi.yaml
openapi: 3.1.0
info:
  title: Project Atlas Synthetic Fraud Defense API
  version: 0.1.0
  description: >-
    Local-only API for Project Atlas, a synthetic, defensive red/blue fraud-model
    evaluation demo. This API must not connect to production systems or real customer data.
servers:
  - url: http://localhost:8000
    description: Local mock API only
security: []
paths:
...
  /synthetic/sample:
    get:
      summary: Return a small public-safe synthetic sample
      operationId: getSyntheticSample
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        "200":
          description: Public-safe synthetic sample records
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyntheticSampleResponse'
  /score:
    post:
      summary: Score one synthetic event
      operationId: scoreEvent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScoreRequest'
      responses:
        "200":
          description: Score and decision action
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScoreResponse'
        "400":
          $ref: '#/components/responses/BadRequest'
        "422":
          $ref: '#/components/responses/ValidationError'
  /batch-score:
    post:
      summary: Score multiple synthetic events
      operationId: batchScoreEvents
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchScoreRequest'
      responses:
        "200":
          description: Batch scores and decision actions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchScoreResponse'
        "400":
          $ref: '#/components/responses/BadRequest'
        "413":
          description: Request exceeds max batch size
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "422":
          $ref: '#/components/responses/ValidationError'
...
  /red-team/search:
    post:
      summary: Run a synthetic red-team search over event histories
      operationId: runRedTeamSearch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RedTeamSearchRequest'
      responses:
        "200":
          description: Search result and model vulnerability cards
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RedTeamSearchResponse'
        "400":
          $ref: '#/components/responses/BadRequest'
        "422":
          $ref: '#/components/responses/ValidationError'
...
components:
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    SyntheticSampleResponse:
      type: object
      properties:
        customers:
          type: array
          items:
            $ref: '#/components/schemas/Customer'
        events:
          type: array
          items:
            $ref: '#/components/schemas/EventRecord'
        features:
          type: array
          items:
            $ref: '#/components/schemas/FeatureVector'
      required: [customers, events, features]
    Customer:
      type: object
      properties:
        customer_id:
          type: string
        customer_segment:
          type: string
        home_region_bucket:
          type: string
        account_age_days:
          type: integer
        normal_login_frequency_30d:
          type: integer
        normal_transfer_frequency_30d:
          type: integer
        synthetic_base_risk:
          type: number
      required: [customer_id, customer_segment, home_region_bucket, account_age_days]
    EventRecord:
      type: object
      properties:
        event_id:
          type: string
        customer_id:
          type: string
        event_type:
          type: string
        event_time_utc:
          type: string
          format: date-time
        device_id:
          type: string
        account_id:
          type: string
        recipient_id:
          type: string
        channel:
          type: string
        amount_bucket:
          type: string
        synthetic_truth_label:
          type: string
          enum: [normal_activity, high_risk_synthetic_activity]
      required: [event_id, customer_id, event_type, event_time_utc, synthetic_truth_label]
    FeatureVector:
      type: object
      properties:
        event_id:
          type: string
        customer_id:
          type: string
        login_count_72h:
          type: integer
        login_count_30d:
          type: integer
        login_velocity_ratio:
          type: number
        challenge_count_72h:
          type: integer
        challenge_pass_ratio_30d:
          type: number
        password_recovery_count_72h:
          type: integer
        device_count_72h:
          type: integer
        current_device_tenure_days:
          type: integer
        geo_consistency_flag:
          type: integer
          enum: [0, 1]
        transfer_count_72h:
          type: integer
        recipient_tenure_days:
          type: integer
        shared_device_degree:
          type: integer
        shared_recipient_degree:
          type: integer
        entity_graph_risk_score:
          type: number
        cash_movement_velocity_score:
          type: number
      required: [event_id, customer_id]
    ScoreRequest:
      type: object
      properties:
        event:
          $ref: '#/components/schemas/EventRecord'
        features:
          $ref: '#/components/schemas/FeatureVector'
        model_version:
          type: string
          default: baseline_v1
        threshold_version:
          type: string
          default: thresholds_v1
      required: [event, features]
    ScoreResponse:
      type: object
      properties:
        event_id:
          type: string
        score:
          type: number
          minimum: 0
          maximum: 1
        decision_action:
          type: string
          enum: [accept, challenge, alert, decline]
        decision_band:
          type: string
        model_version:
          type: string
        threshold_version:
          type: string
        reason_codes:
          type: array
          items:
            type: string
      required: [event_id, score, decision_action, decision_band, model_version, threshold_version, reason_codes]
    BatchScoreRequest:
      type: object
      properties:
        records:
          type: array
          minItems: 1
          maxItems: 5000
          items:
            $ref: '#/components/schemas/ScoreRequest'
      required: [records]
    BatchScoreResponse:
      type: object
      properties:
        scores:
          type: array
          items:
            $ref: '#/components/schemas/ScoreResponse'
        model_version:
          type: string
        threshold_version:
          type: string
      required: [scores]
...
    RedTeamSearchRequest:
      type: object
      properties:
        run_id:
          type: string
        round_id:
          type: integer
        allowed_family_ids:
          type: array
          items:
            type: string
        max_score_queries:
          type: integer
          minimum: 1
        search_methods:
          type: array
          items:
            type: string
            enum: [random, evolutionary, graph_probe]
      required: [run_id, round_id, max_score_queries, search_methods]
    RedTeamSearchResponse:
      type: object
      properties:
        run_id:
          type: string
        round_id:
          type: integer
        valid_high_risk_events_tested:
          type: integer
        accepted_high_risk_events:
          type: integer
        model_miss_rate:
          type: number
        miss_rate_lift_vs_random:
          type: number
        found_adaptive_set_event_ids:
          type: array
          items:
            type: string
        model_vulnerability_cards:
          type: array
          items:
            $ref: '#/components/schemas/ModelVulnerabilityCard'
      required: [run_id, round_id, valid_high_risk_events_tested, accepted_high_risk_events, model_miss_rate]
    ModelVulnerabilityCard:
      type: object
      properties:
        model_vulnerability_id:
          type: string
        round_id:
          type: integer
        family_id:
          type: string
        summary:
          type: string
        valid_high_risk_events_tested:
          type: integer
        accepted_high_risk_events:
          type: integer
        model_miss_rate:
          type: number
        miss_rate_lift_vs_random:
          type: number
        estimated_synthetic_loss_allowed:
          type: number
        affected_decision_action:
          type: string
        safe_cohort_definition:
          type: object
          additionalProperties: true
        recommended_defensive_fix_types:
          type: array
          items:
            type: string
      required: [model_vulnerability_id, round_id, family_id, summary, model_miss_rate]
...
    ErrorResponse:
      type: object
      properties:
        error_code:
          type: string
        message:
          type: string
        safe_to_show:
          type: boolean
      required: [error_code, message, safe_to_show]